Artificial Intelligence/Advanced/LLM (LANGCHAIN GEN AI)/LLM4-Invoice Extractor using Gemini Pro Vision/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb 